New blog post
For the banking industry, the customer interaction through the mobile
channel is predicted to be the prevalent one in the near future. So
banks are demanding topologies that allow to achieve a near-zero
downtime target. Tailoring the appropriate product topology, IBM
MobileFirst Platform Foundation is able to achieve this stringent
target. And to delight your clients, you have to satisfy the unexpressed
requirements too.